remove game level desert, only set 1 game level, update the homepage (game.js)
